+++ src/ts/data/observation.ts
@@ -1,10 +1,10 @@
 /**
  * Rattribute.js
  * 
- * A JavaScript library that generates responsive attribute setters for any HTML element.
+ * A lightweight JavaScript library for automatically changing HTML element attributes based on responsive screen sizes.
  * 
  * @file        observation.ts
- * @version     v1.2.0
+ * @version     v1.3.0
  * @author      Bunoon
  * @license     MIT License
  * @copyright   Bunoon 2026
@@ -18,7 +18,7 @@ import { Is } from "../data/is";
 export namespace Observation {
     let _mutationObserver: MutationObserver = null! as MutationObserver;
 
-    export function setup( configurationOptions: ConfigurationOptions, mutationFunc: Function ) {
+    export function setup( configurationOptions: ConfigurationOptions, mutationFunc: Function ) : void {
         if ( configurationOptions.observationMode ) {
             if ( !Is.defined( _mutationObserver ) ) {
                 _mutationObserver = new MutationObserver( () : void => mutationFunc() );
@@ -33,15 +33,18 @@ export namespace Observation {
             }
             
         } else {
-            _mutationObserver.disconnect();
-            _mutationObserver = null!;
+            disconnect()
         }
     }
 
-    export function destroy( configurationOptions: ConfigurationOptions ) {
+    export function destroy( configurationOptions: ConfigurationOptions ) : void {
         if ( configurationOptions.observationMode && Is.defined( _mutationObserver ) ) {
-            _mutationObserver.disconnect();
-            _mutationObserver = null!;
+            disconnect();
         }
+    }
+
+    function disconnect() : void {
+        _mutationObserver.disconnect();
+        _mutationObserver = null!;
     }
 }